# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LocalSpot Mailer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/active:
    get:
      operationId: getActiveCampaign
      tags: [campaigns]
      summary: Get active campaign with all spots
      responses:
        "200":
          description: Active campaign with spots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithSpots"
        "404":
          description: No active campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /campaigns/by-slug/{slug}:
    get:
      operationId: getCampaignBySlug
      tags: [campaigns]
      summary: Get a published territory campaign by URL slug, with all spots
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published campaign with spots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignWithSpots"
        "404":
          description: No published page for this slug
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/campaigns:
    get:
      operationId: listAdminCampaigns
      tags: [admin]
      summary: List all campaigns with revenue summaries
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCampaignsListResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createCampaign
      tags: [admin]
      summary: Create a new campaign and auto-generate its spot layout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "200":
          description: Campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCampaignDetailResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/campaigns/{id}:
    get:
      operationId: getAdminCampaignById
      tags: [admin]
      summary: Get a specific campaign with all spots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign with spots and stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCampaignDetailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/campaigns/{id}/activate:
    post:
      operationId: activateCampaign
      tags: [admin]
      summary: Mark a campaign active (demotes any other active campaign to completed)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCampaignDetailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/campaigns/{id}/complete:
    post:
      operationId: completeCampaign
      tags: [admin]
      summary: Mark a campaign completed and lock its spots
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCampaignDetailResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        territory:
          type: string
        zipCode:
          type: string
        mailDate:
          type: ["string", "null"]
        homesCount:
          type: integer
        status:
          type: string
          enum: [draft, active, completed]
        slug:
          type: ["string", "null"]
        isPublished:
          type: boolean
        dealerId:
          type: ["integer", "null"]
        mailingSeason:
          type: ["string", "null"]
        mailingMonth:
          type: ["string", "null"]
        cityList:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - territory
        - zipCode
        - homesCount
        - status
        - createdAt
    Spot:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        side:
          type: string
          enum: [front, back]
        size:
          type: string
          enum: [xl, large, medium, small]
        gridArea:
          type: string
        price:
          type: integer
        categoryLock:
          type: ["string", "null"]
        status:
          type: string
          enum: [available, reserved, paid]
        businessName:
          type: ["string", "null"]
        businessCategory:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        adFileUrl:
          type: ["string", "null"]
        adStatus:
          type: ["string", "null"]
        trackingCode:
          type: ["string", "null"]
        templateData:
          type: ["object", "null"]
          additionalProperties: true
        scanCount:
          type: integer
        expiresAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - campaignId
        - side
        - size
        - gridArea
        - price
        - status
        - createdAt
    CampaignWithSpots:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          properties:
            spots:
              type: array
              items:
                $ref: "#/components/schemas/Spot"
          required:
            - spots
...
    AdminSpotRow:
      allOf:
        - $ref: "#/components/schemas/Spot"
        - type: object
          properties:
            isPaid:
              type: boolean
            stripePaymentIntentId:
              type: ["string", "null"]
          required:
            - isPaid
...
    AdminCampaignResponse:
      type: object
      properties:
        campaign:
          $ref: "#/components/schemas/Campaign"
        spots:
          type: array
          items:
            $ref: "#/components/schemas/AdminSpotRow"
        totalRevenue:
          type: integer
        totalSpots:
          type: integer
        paidSpots:
          type: integer
      required:
        - campaign
        - spots
        - totalRevenue
        - totalSpots
        - paidSpots
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        territory:
          type: string
        zipCode:
          type: string
        homesCount:
          type: integer
        mailDate:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, completed]
      required:
        - name
        - territory
        - zipCode
        - homesCount
    CampaignSummary:
      allOf:
        - $ref: "#/components/schemas/Campaign"
        - type: object
          properties:
            totalSpots:
              type: integer
            paidSpots:
              type: integer
            availableSpots:
              type: integer
            totalRevenue:
              type: integer
          required:
            - totalSpots
            - paidSpots
            - availableSpots
            - totalRevenue
    AdminCampaignsListResponse:
      type: object
      properties:
        campaigns:
          type: array
          items:
            $ref: "#/components/schemas/CampaignSummary"
      required:
        - campaigns
...
    AdminCampaignDetailResponse:
      allOf:
        - $ref: "#/components/schemas/AdminCampaignResponse"
        - type: object
          properties:
            availableSpots:
              type: integer
          required:
            - availableSpots